Create USER ID GENERATION.java
Joseph's team has been assigned the task of creating user-ids for all participants of an online gaming competiton. Joseph has designed a process for generating the user-id using the participant's First_name,Last_Name,PIN  code and a number N. The process defined by Joseph is as below-
STEP 1-Compare the lengths of FIRST_Name and Last_name of the participant. The one that is shorter will be called "Smaller Name" and the one that is longer will be called the "Longer name". If  the both First_name and Last_name are of equal length, then the name that appears earlier in alphabetical order will be "Smaller Name" and the name that appears later in alphabetical order will be called the "Longer name".
STEP 2:The user-id should be generated as below-First Letter of the longer name+ Entire word of the smaller name+ digit at position N in the PIN when traversing the PIN from Left to right+ Digit at position N in the PIN when traversing the PIN from right to left.
STEP 3: Toggle the Alphabets of  the user-id generated in step-2.i.e.upper-case alphabets  should become lower-case and lower-case should become upper-case


Test Cases:
I/P: Rajiv Roy 560037 6
O/p:rrOY75

I/p : Manoj Kumar 561327 2
O/p :mkUMAR62

I/p: Kumar Kumud 561327 2
O/p : kkUMAR62